just code cleanup. Mostly just more exact casts.

+++ SELLA/MAIN_SELLA.cpp
@@ -165,12 +165,12 @@ void GenerateCubicTest() {
    const double range = 40;
    int rowcount = 0;
    int colcount = 0;
-   for (iy = -ydel; iy < range; iy += ydel) {
-      y = y0 + iy + ydel;
+   for (iy = (long)(-ydel); iy < range; iy += ydel) {
+      y = y0 + (double)(iy) + ydel;
       colcount = 0;
       ++rowcount;
-      for (iz = -zdel; iz < range; iz += zdel) {
-         z = z0 + iz + zdel;
+      for (iz = (long)(-zdel); iz < range; iz += zdel) {
+         z = z0 + (double)(iz) + zdel;
 
          const S6 redP = ReduceLatticeType("p", s6);
          const S6 redI = ReduceLatticeType("i", s6);
